+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Change GRUB to use partition UUIDs instead of Linux UUIDs.
 - Change Mac hardware workarounds to happen during the first-time setup.
 - Change install to allow using a single HTTP mirror for caching purposes.
+- Add ability to upgrade firmware during the first-time setup.
 
 ## 4.0.0 - The Conversion Update